refactor: moved server info to japrotocol
