+++ jarust_interface/src/lib.rs
@@ -12,7 +12,6 @@ pub mod error;
 pub mod handle_msg;
 pub mod janus_interface;
 pub mod japrotocol;
-pub mod respones;
 pub mod restful;
 pub mod tgenerator;
 pub mod websocket;
